+++ oraclebone/__init__.py
@@ -1,3 +1,3 @@
 """Utilities for the oraclebone command line interface."""
 
-__version__ = "8.1.0"
+__version__ = "8.2.0"